Removed the old test_authenticator.py now that all of its tests have been incorporated into the new tests. Moved the new test file over into test_authenticator.py.
